Set options for redistio

Set options for redistio

Usage

redistio_options(
  theme = "flatly",
  panels = c("elections", "demographics", "integrity", "algorithms"),
  select_color = "purple",
  leaflet_height = "91vh",
  crs = 4326,
  layer_weight = 1.5,
  layer_color = "#000000",
  use_algorithms = TRUE,
  alg_max_districts = 3,
  alg_max_sims = 100,
  alg_counties = NULL,
  ...
)

Arguments

theme

a name of a bootswatch preset theme or other bslib::bs_theme() object

panels

which panels to display in the app, 'draw' is always shown.

select_color

a color to use for highlighting selected districts

leaflet_height

height to pass to leaflet::leafletOutput()

crs

a coordinate reference system to use in leaflet::leaflet()

layer_weight

a stroke width to use for layers in leaflet::leaflet()

layer_color

a color to use for layers in leaflet::leaflet()

use_algorithms

whether to use redistricting simulation algorithms

alg_max_districts

maximum number of districts to use in algorithms

alg_max_sims

maximum number of simulations to use in algorithms

alg_counties

name of county column in sf object to use in algorithms

...

additional arguments (currently ignored)

Value

a list

Examples

redistio_options()
#> $theme
#> [1] "flatly"
#> 
#> $panels
#> [1] "elections"    "demographics" "integrity"    "algorithms"  
#> 
#> $select_color
#> [1] "purple"
#> 
#> $leaflet_height
#> [1] "91vh"
#> 
#> $crs
#> [1] 4326
#> 
#> $layer_weight
#> [1] 1.5
#> 
#> $layer_color
#> [1] "#000000"
#> 
#> $use_algorithms
#> [1] TRUE
#> 
#> $alg_max_districts
#> [1] 3
#> 
#> $alg_max_sims
#> [1] 100
#> 
#> $alg_counties
#> NULL
#>